Refactor CommandsHelper class and Help module
The 'Help' namespace was renamed to `Support`
and the 'help' directory was renamed to 'support'.

Since 'CommandHelper' was a confusing utility name
(it just handles commands help), it was renamed to 'Help'
and moved to 'commands/support' directory. Now its full name is
'SlackRubyBot::Commands::Support::Help'.
