+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2018-03-06 16:06:41 +0100 using RuboCop version 0.51.0.
+# on 2018-03-29 16:47:34 +0200 using RuboCop version 0.51.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -43,11 +43,11 @@ Metrics/ClassLength:
 Metrics/CyclomaticComplexity:
   Max: 16
 
-# Offense count: 248
+# Offense count: 249
 # Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
 Metrics/LineLength:
-  Max: 147
+  Max: 151
 
 # Offense count: 11
 # Configuration parameters: CountComments.
