+++ lib/slack-ruby-bot.rb
@@ -19,7 +19,6 @@ module SlackRubyBot
 end
 
 require 'slack-ruby-client'
-require 'slack-ruby-bot/support/commands_helper'
 require 'slack-ruby-bot/commands'
 require 'slack-ruby-bot/client'
 require 'slack-ruby-bot/server'
